filter_ecs: reduce error/warn messages when the filter processes logs not from a task
The filter is built primarily for a daemon deployment mode where Fluent Bit
runs once per node/instance and collects all log files on that host. The
filter can attach metadata to these logs. But- what if there is a
container running on the host that is not part of an ECS Task?
Json-file log driver files are written to disk with only the
container ID to distinguish where they came from. So the
daemon would still collect logs from containers not part of a task
and there is no easy way to ignore those logs. Without this patch,
in that case, the Fluent Bit output would be spammed with constant
errors from this filter. This patch suppresses failures after
2 failed attempts.

Signed-off-by: Wesley Pettit <wppttt@amazon.com>