out_s3: locking chunks is no longer needed for async http
S3 now always uses sync IO. In the future, we will
switch back to async with the concept of a daemon
coroutine that is the only thing that can send data.
Either way, we don't need to mark that chunks are
actively being sent with the locked bool

A subsequent commit will refactor this code to
use locked for a similar but slightly different
purpose. Locked will mean a chunk is ready/done
and there should be no new appends to it.
Locked chunks will be added to a queue for sending.
Unlocked chunks are ready for new flushes to buffer
new data.

Signed-off-by: Wesley Pettit <wppttt@amazon.com>